logcat: Initialize Diag only if it is available
The issue is here is that initialising Diag with insufficient privilege
causes a message to be written to standard output.  This is then read
by the test app and appears as a test failure.

The fix is to supress the initialisation of Diag if the device file
that Diag uses is not readable by the calling app.