system/core: Fix the issue of not able to unmount SD card from UI option
SD card is not able to unmount when 'Unmount SD card' option
is selected in UI Settings. When this option is selected UI,
send Eject_media command to volume daemon. volume daemon
checks the volume whose mount point is sdcard. It was unable
to find the volume even though it is present in the list.

Fix the issue by correcting volume name search algorithm.
It also fixes the auto mount issue on 7x30.

CRs-fixed: 219070