init: Correct the implementation of do_devwait function.
When the device nodes are already available do_devwait should immediately
return, which however is not happening and the control is stuck in this
loop forever. The issue is found to be the nested if-else construct that
is now corrected in this patch. Else block gets associated with the last
unmatched if block and in this case the same has caused the else block to
get associated with the wrong if block.