init: Correct the do_devwait function to properly timeout.
The variable that keeps track of the timeout is not being handled properly.
It is not getting decermented as expected and the control waits in this
function do_devwait forever in case of the device handles that are never
expected to appear. For example watiting for /dev/mmcblock0p14 (on a card
that contains only 13 partitions) is expected to timeout, but, it does not
happen because we simply do a continue when poll timesout and returns 0.
This patch properly decrements this timeout variable to ensure that we exit
this function properly.