vold: Add support for user-specified SD card partition
New property "emmc.sdcard.partition" indicates the
partition of SD card to mount for /sdcard.  This can
be a partition in the extended partition table.  Property
is a number starting at 1 (e.g. "1" == mmcblkXp1).

This feature is intended to be used on eMMC devices with
no external SD card, and thus must use one of the eMMC
partitions for /sdcard.

When the SD card is formatted via vold, the partition will
auto-remount.

Change-Id: Ic5c24d10d7a03ca7496158c0853749b7541e4e85