rootdir: Add battery charging as part of new boot-pause section
If we are booting up due to power charger connected to target, we do
not want to go to full Android framework boot -- instead we stay in
boot-pause section in battery_charging until user indicates to
continue to boot.  The user indication logic can be a simple key
press and is part of battery_charging code.

Change-Id: I533a33e5b11bad871eed40c6e60b9b19e7902696