system: init: Fail early when missing firmware
Currently, when firmware cannot be found to satisfy a
firmware_request() from the kernel the init process continues
without notifying the kernel in anyway. Instead of doing this,
tell the kernel that the firmware is being looked for, and then
fail the loading process if the firmware couldn't be found. This
should speed up callers in the kernel who are waiting on firmware
that will never be found.

Change-Id: I5dc62d0d6f3bc1062815fa27feb60603b3608ae4
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>