Implement Autotools for libcutils
This implementation more closely builds what Android builds for libcutils.
However, it's now possible that both memset32.S and memory.c will be
built and linked, leading to multiple definitions of a couple symbols. This
is resolved by striping the conflicting symbols from memory.c before the
library is linked.

Change-Id: Iff89b706206f4f277dfaa8ae2e44f056ff920fe2