Change setting of /system/bin/sh execute /system/bin/btwlancoex as non-root.
By default, the owner of /system/bin/btwlancoex, which is called from
/system/etc/init.qcom.coex.sh, is set to be root.
To satisfy Google's requirement that /system/bin/sh executing
/system/bin/btwlancoex should not act as root,  we have changed permission
of /system/bin/btwlancoex to execute as the owner of the file (i.e. root)
and the user of .sh process. executing /system/etc/init.qcom.coex.sh is shell.

Change-Id: I2df92f668c8d4de4563f8cddebca9156ccd9436d
CRs-fixed: 275304