Bluetooth: BT 3.0+HS (AMP) should be disabled by default
AMP will be enabled only for product lines that have the feature.
This is necessary because of the side-effect that the WLAN driver is
loaded when Bluetooth is turned on if AMP is enabled.

Signed-off-by: Peter Krystad <pkrystad@codeaurora.org>
CRs-fixed: 287992
(cherry picked from commit f6d4f512fd41e3f8b0d9962bb06f44b774ad7f40)

Change-Id: I5deace6050b7cc781422c043bf02c81b7d2f1791