init: Enhance support for loading the wcnss module
It is anticipated that the current wcnsswlan kernel module will be
undergoing some transformations.  Possible transformations include
renaming the module (i.e. from wcnsswlan to wcnss) and changing the
build options so that the module is built-in to the kernel (and
hence it would no longer need to be dynamicially loaded).  In order
to support all of these possibilities we enhance the msm8960 device
init processing to dynamicially determine how to initialize the
wcnss device.

Change-Id: Iba91cea0f465fc960798f9bdf251e7b695a25ab2