ADB: Update fbinfo according to endianness
If the framebuffer is populated in big-endian fashion and is 32 bits per pixel,
screen capture doesn't work correctly. This is because DDMS screen capture
expects 32bpp buffers to be in little-endian format.
So, if we see a big-endian 32bpp buffer, change it to little endian so that
DDMS screen capture shows the correct output.

Change-Id: I95d64605ed5f047a6f4946a8235f0fec9df553f1
CRs-fixed: 296687