ADB: Update fbinfo according to endianness
If the framebuffer is populated in big-endian fashion and is 32 bits per pixel,
screen capture doesn't work correctly. This is because DDMS screen capture
expects 32bpp buffers to be in little-endian format.
So, if we see a big-endian 32bpp buffer, change it to little endian so that
DDMS screen capture shows the correct output.

CRs-fixed: 296687
(cherry picked from commit 9aa8c7a9f3513d5eaaef2e44b3632a658bcf038e)

Change-Id: I78180fd5749a74972f25f9575f278b38c517385b