mkbootimg: Add new mandatory ramdisk offset parameter
We can't assume a constant offset for the ramdisk start address
for all targets. This varies based on the board memory size and
size of the uncompressed kernel.
This is the offset from the kernel address.
This currently supports booting an uncompressed kernel only.
The base parameter should be specified before the ramdisk_offset.

Change-Id: Ia9be0ef3d2b76cc476aa75b77ee116fb6fd6dcda