Added reboot and shutdown scripts to powerapp
As glibc does not seem to support a working reboot function, the
powerkey app has been adapted to allow to invoke the correct
system call.  In order to be able to shutdown or reboot the target
two init scripts have been added to force the target to shutdown
or reboot when reaching the appropriate run level.

Change-Id: Ie3927dc842a12882255826ea0a5ca6f8c5bc293f