powerapp: Add support for reboot to accept an argument
The kernel uses the argument passed to reboot to set the
restart reason in shared memory. This restart reason is read
by the apps bootloader to perform various operations such as
waiting explicitly in fastboot to process commands.
This is one of the things required to add support for
reboot bootloader.

Change-Id: I8e030c5c6b49cb2960d126793eaad154d754c4e0