init: Create USB host bus directory during boot up
Currently ueventd process is creating /dev/bus/usb directory as
part of the first USB device (root hub of a bus) discovery. That
means if USB host bus is not added during kernel initialization,
this directory will not be created at boot time.  Libusbhost tries
to monitor this directory during boot-up and failing.  Due to this
UsbHostManager never get to know about connected USB devices and no
application is notified about the same.

It is possible that USB host bus is added dynamically when Micro-A cable
or dock is connected.  Hence Create /dev/bus/usb once for all during init
and avoid creating the same for every USB device connection.

Change-Id: Icc7ec71d9452853602a653a4b9dfb79e4ffdaeb4