mkbootimg: add flag to put kernel after initrd
If a compressed kernel image is used, it does not make sense to
load the kernel at the same location the uncompressed kernel image
will reside at. Doing so forces the decompressor/compressed image
to move itself somewhere else.

Add a flag indicating that a compressed kernel image is being used.
If this flag is set the compressed image will be loaded immediately
after the ramdisk. If there is no ramdisk, the kernel will be loaded
where the ramdisk would have been.

Change-Id: I508df1ef40d4d4e63f0d2bd004742e56ba78fe17