quipc_main: changed order of groups to have gps first
This is to resolve the problem when IGSD (a module of
quipc_main) tries to call icm_init(). This function
creates a pipe to which the HAL will reply when Wifi
has either been brought up or failed to come up. This
pipe resides in gpsone_d folder which has system owner
and gps group.  If quipc_main does not have gps group
first, it crashes, so this change is mandatory.

Change-Id: I84ead837ed290997638fe198ad070bf232221f85