init: Change ownership of sampling_down_factor file to system
Changing ownership to system allows CpuGovernorService
to modify sampling_down_factor based on votes from
applications such as the camcorder which need to set
sampling_down factor to 1 in order to reduce the amount
of time for which the CPU runs at the maximum frequency.
Without the ownership change, CpuGovernorService would
not be able to write to the sampling_down_factor file.
Also fixing io_is_busy so all targets will be able to
write to it.

Change-Id: Ic2b1abd194db18efac7ca624392ba1904cbe4762