init: qcom: Move system partition remount stage to early-boot
During initialization the system partition needs to be
remounted as read-write in order for changes to be made
to the /system partition (e.g. creating target-specific
symlinks, etc.).This change moves these initializations
to the early-boot stage.

TODO: Init sequence needs to be redone for cleanup
and boot up speed.

CRs-Fixed: 397912
Change-Id: I428dc1633c7315861f9fb695305c8d481168c38c