adb: Fix Windows ADB stdout handling when spawning server
Mark stdout handle as non-inheritable before spawning server
process on Windows.  This fix the issue where parent process
that waits for adb to complete based on stdout to hang
forever when ADB needs to start the server.

  e.g.  abc=`adb devices`

The above command previously does not return if ADB server is
started as a result of the ADB command.  This is because the
server process inherits the stdout handle causing the command
to only return when the server portion exits.

CRs-Fixed: 406617
Change-Id: Ia2f4107aba4eb119e788a96751bac5e171d5e60a