Add new USB composition PID 0x9063 - RNDIS : ECM : MBIM
This is a composition consisting of multiple configurations
intended to support plug-in to a host with "zero installation".

Configuration 1 - RNDIS (Win7 and older hosts)
Configuration 2 - ECM (Mac OS X and Linux hosts)
Configuration 3 - MBIM (Win8 and newer hosts)

Corresponding changes in the MBIM kernel driver have been added to
return Microsoft-specific OS descriptors that allow Windows 8 hosts
to dynamically select the third configuration when the device is
enumerated.

CRs-fixed: 410347
Change-Id: I884489432b5fe82e5185342d527c930d9b415a37