platform/system/core: Update powerapp to prevent bogus resume
In the linux kernel when device suspend happens
all userspace processes are frozen. When the
powerkey is used to trigger an interrupt to resume
the device after suspend, the userspace has already
been unfrozen and has resumed as the powerkey interrupt
is a wakeup interrupt. The logic in powerapp
resume doesnt come into effect. Hence remove this
portion of the code.

Change-Id: I088f3cf730bfa6528f1a78b51ae2a3e30de5724c
Signed-off-by: Chandra Ramachandran <cramacha@codeaurora.org>