Fix data call failure issue in USB composition 902E
In USB compostion 902E, both RMNET and RNDIS functions are enabled in
BAM2BAM mode while interfacing with the A2 core. However, the A2 core is
capable of working one BAM2BAM channel only, and therefore there is a
collision between RNDIS and RMNET over the A2 BAM2BAM path. The solution
impleneted for this issue is to change the RNDIS function to run in
non-BAM mode in this composition.

CRs-fixed: 498219
Change-Id: I50c5f4358bd7ab9e5d6130f19a0cde5d91895d7a
Signed-off-by: Daniel Segal <dsegal@codeaurora.org>