system: core: usb: Add USB debug script
Script for enable or disable debug messages in USB related files.
this script allows the user to chose the files in which should
enable or disable debug messages.
and also allows the user to save the debug configuration
for the next power cycle,this in order to allow early boot
debug information.

Change-Id: I59a57a9c3bced4c26c9f9cc2a65cd8cea2494f08
Signed-off-by: Nadine Toledano <nadinet@codeaurora.org>