charger: add time check to wait for the right alarm
If power off alarm was set by user, alarm thread in power
off charging will set and wait for the right time to reboot
the phone. During the process of waiting, the rtc time and
our alarm time are compared to determine whether need to
exit from the waiting loop.

Change-Id: I9139f1e536db5eb8107e2963c846621b09428d15