Add a configuration node for support concurrent SoftAP and DUN
To support SoftAP and DUn concurrently we need the option
to switch from TTY to SMD mode and vice versa. We allow
this option by enabling the following:
echo Y > /usr/bin/usb/softap_w_dun
After this it becomes possible to issue the special ioctl to
switch from SMD to TTY, and also poll on the special
device attributes to get the current status of the cable.

Change-Id: I474014108ca0d2409f5fd5b96d577b450a179e65
Signed-off-by: Lena Salman <esalman@codeaurora.org>