usb: composition: Fix race condition between starting and stopping adbd
With adb reboot scenario, it has been observed that start-stop-daemon
is racing with starting and stopping of adbd daemon on bootup as adbd
is being started 2 times and once it is being stopped. start-stop-deamon
forks the process and it is not blocking call, it results into where
started adbd daemon is being stopped cauing no usb enuermation. Hence
instead of stopping adbd daemon using start-stop-daemon, kills adbd
daemon to over come this race condition.

Change-Id: I2f09e3a0db94a3bc307b7dbfc2404904087fd103
Signed-off-by: Mayank Rana <mrana@codeaurora.org>