usb: compositions: Fix context switch to f_fs (ADB) composition
On composition switch to f_fs (ADB) composition, we need to re-start
the ADB daemon in order for the f_fs bind function to re-occur.

The problem is that the ADB daemon was not actually re-started since when
doing '/etc/init.d/adbd start' command, it was notified that the daemon
is already running and therefore the start command was ignored.
In order to avoid that, we simply add a stop command before we kill the
daemon on the re-start sequence.

Change-Id: I38a98602769f9703eb17746b2a0bb7357e5bba6d
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>