usb: composition: Fix ADB daemon kill command
With adb reboot scenario, it has been observed that start-stop-daemon
is racing with starting and stopping of adbd daemon on bootup as adbd
is being started 2 times and once it is being stopped. start-stop-deamon
forks the process and it is not blocking call, it results into where
started adbd daemon is being stopped cauing no usb enuermation. Hence
instead of stopping adbd daemon using start-stop-daemon, kills adbd
daemon to over come this race condition.

This patch fix the pkill command of the adbd, to be more generic using
`which adb`. This fix an issue, that due to adbd location change, the
pkill command stopped from being effective.

Change-Id: Ie28963786136e2330ef792140bc49f94c594d727
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>
Signed-off-by: Mayank Rana <mrana@codeaurora.org>