system: core: usb: Stop ADB correctly on composition change
It is needed to stop ADB on composition change.
Especially it's critical when changing from composition no
containing adb to one that does. A change like this without
stop to adb is causing the target to crash.
This was solved, however the command to stop the adb
with time ceased to actually stop this.
This commit reverts this functionality to working state.

Change-Id: I1e932974da24a83f86d54aa74667dfa5c2659027
Signed-off-by: Lena Salman <esalman@codeaurora.org>