usb: compositions: kill the earliest adbd instance if running
Upon boot, the composition switch script is called
with a delay of 0ms prior to enabling adbd. This causes
a race with the pkill directive in the beginning of the
script. In order to avoid the race, add the -o flag to only
kill the earliest instance of adbd. Furthermore, only kill
adbd if it is running, in order to avoid a situation where
the earliest instance will, once again, be a race between
pkill and adbd start.

Change-Id: I460a3e7e97616b736b4bdee0d6f3ab754f50d88f
Signed-off-by: Dov Levenglick <dovl@codeaurora.org>