usb: composition: don't force different functions for USB and HSIC
Changes the behavior of the composition switching script to
allow the same functions to potentially run on both USB and
HSIC ports. This is required to support existing use-cases
where the USB port has a composition of RNDIS:ECM and
the HSIC port has ECM running. Such a case should be
supported so long as the USB host chooses RNDIS over ECM.
The change clearly warns the end user of the danger, yet
leaves the final decision up to the user; rather than
not supporting the composition change.

Change-Id: If784400415a189673017e224ef3eed0b06198803
Signed-off-by: Dov Levenglick <dovl@codeaurora.org>