system: core: usb: Initialize HSIC before HSUSB
Kernel adb driver (f_fs) disables USB enumeration if
adbd is not running but ADB is present in composition.
If HSUSB is started first then ADB is enabled and adbd
is started. Later as part of HSIC composition rules
adbd is killed which disables USB enumeration on HSUSB
as well. Change this sequence so that HSIC is enabled
first and later HSUSB script can enable and start adbd.
Also add delay between HSIC and HSUSB initialization
to avoid both corrupting PID/VID in device_desc which
is shared for both HSUSB and HSIC.

Change-Id: I9b1b8eda7a94aef35146532adf57bfb8f67f7f70
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>